feat(bet4-drain18): extract marketing-ops tools into a scoped ToolPack (EP-8DC217EB BET-4)
Move the 16 marketing-ops/autopilot MCP tools (get_marketing_summary,
… analyze_seo_opportunity) out of the mcp-tools.ts executeTool switch into
apps/web/lib/mcp/packs/marketing-ops-pack.ts, registered via pack-registry.ts.
Behavior-identical dispatch; baseline shrinks by 16; module-size retightened.

Process-Spine-Decision: per 2026-07-07-vertical-integration-inward-plan.md §4 BET-4 (BI-2B7EE073), kernel drain-to-zero (HIGH, margin 2.257) — extract a clean cohesive domain, ratchet freezes the inline set against regrowth.
