+++ scripts/mcp-tool-pack-baseline.json
@@ -10,7 +10,6 @@
   "analyze_brand_document",
   "analyze_public_website_branding",
   "analyze_reusability",
-  "analyze_seo_opportunity",
   "approve_decomposition",
   "assess_archetype_refinement",
   "assess_complexity",
@@ -25,16 +24,12 @@
   "create_build_epic",
   "create_digital_product",
   "create_epic",
-  "create_marketing_asset_task",
-  "create_marketing_automation_candidate",
-  "create_marketing_campaign_brief",
   "create_portal_pr",
   "create_release_bundle",
   "deploy_feature",
   "describe_model",
   "diagnose_sandbox",
   "dpf_test_kernel_refuse_probe",
-  "draft_marketing_asset",
   "drive_browser_task",
   "edit_sandbox_file",
   "evaluate_page",
@@ -56,7 +51,6 @@
   "get_child_threads",
   "get_code_graph_freshness",
   "get_finance_period_summary",
-  "get_marketing_summary",
   "get_my_coworker_profile",
   "get_next_recommended_work",
   "get_release_status",
@@ -74,8 +68,6 @@
   "list_patch_posture",
   "list_project_directory",
   "list_sandbox_files",
-  "place_linkedin_ad",
-  "plan_upcoming_marketing_drafts",
   "prefill_onboarding_wizard",
   "principle_decide",
   "process_backlog_for_build_studio",
@@ -86,7 +78,6 @@
   "propose_improvement",
   "propose_skill_improvement",
   "provision_build_engine",
-  "publish_to_linkedin",
   "query_backlog",
   "read_codebase_manifest",
   "read_project_file",
@@ -97,9 +88,7 @@
   "record_external_development_evidence",
   "record_functional_failure_evidence",
   "record_local_integration_result",
-  "record_marketing_kpi_checkpoint",
   "recover_sandbox",
-  "refresh_channel_kpis",
   "register_digital_product_from_build",
   "request_coworker",
   "resolve_model_selection",
@@ -114,7 +103,6 @@
   "run_ux_test",
   "saveBuildEvidence",
   "save_build_notes",
-  "save_marketing_review",
   "save_phase_handoff",
   "schedule_promotion",
   "schedule_release_bundle",
@@ -126,9 +114,7 @@
   "search_sandbox",
   "search_specs_and_plans",
   "search_tool_marketplace",
-  "send_marketing_email",
   "set_change_disposition",
-  "set_marketing_autopilot_policy",
   "setup_email",
   "size_backlog_item",
   "spawn_work_thread",
@@ -138,10 +124,8 @@
   "start_scout_research",
   "submit_coworker_capability_need",
   "submit_feedback",
-  "suggest_campaign_ideas",
   "suggest_taxonomy_placement",
   "summon_coworker",
-  "tick_marketing_scheduler",
   "trace_code_surface",
   "triage_backlog_item",
   "trigger_contributor_inventory_sync",
